add teacher table view
