+++ README.md
@@ -1,1 +1,5 @@
-LMS - Group A
+# LMS - Group A
+## Install 
+#### npm -install
+#### user: string@gmail.com 
+#### password: asdasd
