+++ src/pages/DashBoard/Dashboard.jsx
@@ -65,7 +65,8 @@ const styles = theme => ({
         display: 'none'
     },
     title: {
-        flexGrow: 1
+        flexGrow: 1,
+        useNextVariants: true,
     },
     drawerPaper: {
         position: 'relative',
